Remove 'all__vary_rounds' parameter since number of rounds is fixed.
As of passlib 1.7 we were getting a deprecation warning seen below:

passlib/utils/handlers.py:1684: PasslibConfigWarning: The 'vary_rounds'
option is deprecated as of Passlib 1.7, and will be removed in Passlib 2.0

Removing this line eliminates the deprecation warning. In addition,
we didn't really want to vary the number of rounds in the first place.
The number of rounds is precisely configured and should always be set
to 1000.